+++ pyleco/utils/pipe_handler.py
@@ -26,6 +26,7 @@ from __future__ import annotations
 from enum import Enum
 from threading import get_ident, Condition
 from typing import Any, Callable, Optional, Union
+from warnings import warn
 
 import zmq
 
@@ -60,7 +61,6 @@ class LockedMessageBuffer(MessageBuffer):
     check, whether that message fits the conversation_id.
     This is repeated until the suiting response is found or a limit is reached.
     """
-    _result: Message
 
     def __init__(self, **kwargs) -> None:
         super().__init__(**kwargs)
@@ -80,34 +80,31 @@ class LockedMessageBuffer(MessageBuffer):
             super().add_message(message)
             self._buffer_lock.notify_all()
 
-    # def add_response_message(self, message: Message) -> bool:
-    #     """Add a message to the buffer, if it is a requested response.
+    def add_response_message(self, message: Message) -> bool:
+        """Add a message to the buffer, if it is a requested response.
 
-    #     :return: whether the message was added to the buffer.
-    #     """
-    #     with self._buffer_lock:
-    #         if message.conversation_id in self._requested_ids:
-    #             self._messages.append(message)
-    #             self._requested_ids.remove(message.conversation_id)
-    #             self._buffer_lock.notify_all()
-    #             return True
-    #         else:
-    #             return False
+        .. deprecated:: 0.3
+            Use :meth:`add_message` instead.
 
-    def retrieve_message(self, conversation_id: Optional[bytes]) -> Optional[Message]:
+        :return: whether the message was added to the buffer.
+        """
+        warn("`add_response_message` is deprecated, use `add_message` instead.", FutureWarning)
+        if self.is_cid_requested(message.conversation_id):
+            self.add_message(message)
+            return True
+        else:
+            return False
+
+    def retrieve_message(self, conversation_id: Optional[bytes] = None) -> Optional[Message]:
         """Retrieve the requested message or the next free one for `None`."""
         with self._buffer_lock:
             return super().retrieve_message(conversation_id=conversation_id)
 
-    def _predicate_generator(self, conversation_id: bytes) -> Callable[[], bool]:
-        def check_message_in_buffer() -> bool:
-            for (i, message) in enumerate(self._messages):
-                if message.conversation_id == conversation_id:
-                    del self._messages[i]
-                    self._requested_ids.discard(conversation_id)
-                    self._result = message
-                    return True
-            return False
+    def _predicate_generator(self, conversation_id: bytes) -> Callable[[], Optional[Message]]:
+        def check_message_in_buffer() -> Optional[Message]:
+            # use verbose super call, as it is necessary for the generated method
+            return super(LockedMessageBuffer, self).retrieve_message(
+                conversation_id=conversation_id)
         return check_message_in_buffer
 
     def wait_for_message(self, conversation_id: bytes, timeout: float = 1) -> Message:
@@ -117,11 +114,11 @@ class LockedMessageBuffer(MessageBuffer):
         :param timeout: Timeout in seconds.
         """
         with self._buffer_lock:
-            found = self._buffer_lock.wait_for(
+            result = self._buffer_lock.wait_for(
                 self._predicate_generator(conversation_id=conversation_id),
                 timeout=timeout)
-            if found:
-                return self._result
+            if result:
+                return result
         # No result found:
         raise TimeoutError("Reading timed out.")
 
@@ -135,7 +132,7 @@ class CommunicatorPipe(CommunicatorProtocol, SubscriberProtocol):
     def __init__(self,
                  handler: ExtendedMessageHandler,
                  pipe_port: int,
-                 buffer: LockedMessageBuffer,
+                 message_buffer: LockedMessageBuffer,
                  context: Optional[zmq.Context] = None,
                  timeout: float = 1,
                  **kwargs):
@@ -145,7 +142,8 @@ class CommunicatorPipe(CommunicatorProtocol, SubscriberProtocol):
         self.socket: zmq.Socket = context.socket(zmq.PAIR)
         self.socket.connect(f"inproc://listenerPipe:{pipe_port}")
         self.rpc_generator = handler.rpc_generator
-        self.buffer = buffer
+        self.message_buffer = message_buffer
+        self.buffer = self.message_buffer  # for backward compatibility
         self.timeout = timeout
 
     # CommunicatorProtocol
@@ -179,12 +177,13 @@ class CommunicatorPipe(CommunicatorProtocol, SubscriberProtocol):
                      ) -> Message:
         if conversation_id is None:
             raise ValueError("You have to request a message with its conversation_id.")
-        return self.buffer.wait_for_message(conversation_id=conversation_id,
-                                            timeout=self.timeout if timeout is None else timeout,
-                                            )
+        return self.message_buffer.wait_for_message(
+            conversation_id=conversation_id,
+            timeout=self.timeout if timeout is None else timeout,
+        )
 
     def ask_message(self, message: Message, timeout: Optional[float] = None) -> Message:
-        self.buffer.add_conversation_id(message.conversation_id)
+        self.message_buffer.add_conversation_id(message.conversation_id)
         self.send_message(message=message)
         return self.read_message(conversation_id=message.conversation_id, timeout=timeout)
 
@@ -211,7 +210,7 @@ class CommunicatorPipe(CommunicatorProtocol, SubscriberProtocol):
     def _send_handler(self, method: str, **kwargs) -> bytes:
         cid = generate_conversation_id()
         message_string = self.rpc_generator.build_request_str(method=method, **kwargs)
-        self.buffer.add_conversation_id(cid)
+        self.message_buffer.add_conversation_id(cid)
         self._send_pipe_message(PipeCommands.LOCAL_COMMAND, cid, message_string.encode())
         return cid
 
@@ -242,6 +241,7 @@ class PipeHandler(ExtendedMessageHandler):
 
     :attr name_changing_methods: List of methods which are called, whenever the full_name changes.
     """
+    message_buffer: LockedMessageBuffer
     _communicators: dict[int, CommunicatorPipe]
     _on_name_change_methods: set[Callable[[str], None]] = set()
 
@@ -251,7 +251,7 @@ class PipeHandler(ExtendedMessageHandler):
         self.internal_pipe: zmq.Socket = context.socket(zmq.PULL)
         self.pipe_port = self.internal_pipe.bind_to_random_port("inproc://listenerPipe",
                                                                 min_port=12345)
-        self.buffer = LockedMessageBuffer()
+        self.message_buffer = LockedMessageBuffer()
         self._communicators = {}
 
     def close(self) -> None:
@@ -323,8 +323,8 @@ class PipeHandler(ExtendedMessageHandler):
     # Local messages
     def handle_local_request(self, conversation_id: bytes, rpc: bytes) -> None:
         result = self.rpc.process_request(data=rpc)
-        self.buffer.add_conversation_id(conversation_id)
-        self.buffer.add_message(
+        self.message_buffer.add_conversation_id(conversation_id)
+        self.message_buffer.add_message(
             Message(
                 "comm",
                 sender="ego",
@@ -337,7 +337,7 @@ class PipeHandler(ExtendedMessageHandler):
     # Thread safe methods for access from other threads
     def create_communicator(self, **kwargs) -> CommunicatorPipe:
         """Create a communicator wherever you want to access the pipe handler."""
-        com = CommunicatorPipe(buffer=self.buffer, pipe_port=self.pipe_port,
+        com = CommunicatorPipe(message_buffer=self.message_buffer, pipe_port=self.pipe_port,
                                handler=self,
                                **kwargs)
         self._communicators[get_ident()] = com